Creating caesar.c
Here, we create an encrypted text example that is firstly planned and applied with Caesar in the Roman era. Here, we get a letter of the string and we change the letter with the next letter that comes right after it in the alphabet. For exp; the string is "hello". In the alphabet, "I" comes right after "H". So, the encrypted text starts with "I". Then, rest of the string continues with this calculation. Here, we use CS50 header file that is called "cs50.h".